+++ tests/Core/Serialization/JSON/Normalizers/LicenseRepositoryNormalizerTest.php
@@ -24,7 +24,9 @@ declare(strict_types=1);
 namespace CycloneDX\Tests\Core\Serialization\JSON\Normalizers;
 
 use CycloneDX\Core\Collections\LicenseRepository;
+use CycloneDX\Core\Models\License\LicenseExpression;
 use CycloneDX\Core\Models\License\NamedLicense;
+use CycloneDX\Core\Models\License\SpdxLicense;
 use CycloneDX\Core\Serialization\JSON\_BaseNormalizer;
 use CycloneDX\Core\Serialization\JSON\NormalizerFactory;
 use CycloneDX\Core\Serialization\JSON\Normalizers\LicenseNormalizer;
@@ -68,12 +70,38 @@ class LicenseRepositoryNormalizerTest extends TestCase
             'getItems' => [$license],
         ]);
 
-        $licenseNormalizer->expects(self::once())->method('normalize')
+        $licenseNormalizer->method('normalize')
             ->with($license)
             ->willReturn(['FakeLicense' => true]);
 
         $actual = $normalizer->normalize($licenses);
 
         self::assertSame([['FakeLicense' => true]], $actual);
+    }
+
+    public function testNormalizePreferExpression(): void
+    {
+        $spec = $this->createStub(Spec::class);
+        $licenseNormalizer = $this->createMock(LicenseNormalizer::class);
+        $factory = $this->createConfiguredMock(NormalizerFactory::class, [
+            'getSpec' => $spec,
+            'makeForLicense' => $licenseNormalizer,
+        ]);
+        $normalizer = new LicenseRepositoryNormalizer($factory);
+        $licenseNamed = $this->createStub(NamedLicense::class);
+        $licenseSpdx = $this->createStub(SpdxLicense::class);
+        $licenseExpression = $this->createStub(LicenseExpression::class);
+        $licenses = $this->createConfiguredMock(LicenseRepository::class, [
+            'count' => 1,
+            'getItems' => [$licenseSpdx, $licenseNamed, $licenseExpression],
+        ]);
+
+        $licenseNormalizer->method('normalize')
+            ->with($licenseExpression)
+            ->willReturn(['FakeLicenseExpression' => true]);
+
+        $actual = $normalizer->normalize($licenses);
+
+        self::assertSame([['FakeLicenseExpression' => true]], $actual);
     }
 }